requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
gunicorn==23.0.0
openpyxl==3.1.5
pandas==2.2.3
python-dotenv==1.0.1
